Move dependency of chapter outside of fingerprint manager
